Software_STM/requirements.txt
pyserial
matplotlib
PyQt5